pyproject.toml
[tool.poetry]
name = "events-collector"
version = "0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "events_collector"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0.4"
environs = "9.5.0"
psycopg2-binary = "2.9.5"
requests = "2.28.1"
djangorestframework = "3.13.1"
beautifulsoup4 = "4.11.1"
stix2-elevator = "4.1.7"
flatdict = "4.0.1"
dagster = "1.0.16"
kafka-python = "2.0.2"
dagit = "1.0.16"
django-filter = "22.1"
aiohttp = "3.8.3"
flake8 = "5.0.4"
django-configurations = "2.4"
django-apscheduler = "0.6.2"
loguru = "0.6.0"
gunicorn = "20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
